requirements.txt
torch
torchvision
timm
pillow
flask
